refactoring schedule
