+++ bot/handlers/schedule/__init__.py
@@ -1,0 +1,2 @@
+from bot.handlers.schedule import days, week, groups_templates, educators, \
+    sending
